set up graphql apollo client
