+++ backend/schema.ts
@@ -1,0 +1,8 @@
+import { makeExecutableSchema } from "graphql-tools";
+import { typeDefs } from "./type-defs";
+import { resolvers } from "./resolvers";
+
+export const schema = makeExecutableSchema({
+  typeDefs,
+  resolvers,
+});
