+++ package.json
@@ -10,8 +10,10 @@
     "codegen": "graphql-codegen --config codegen.yml"
   },
   "dependencies": {
+    "@apollo/client": "^3.4.11",
     "apollo-server-micro": "^2.17.0",
     "graphql": "^15.3.0",
+    "graphql-tools": "^8.2.0",
     "next": "9.5.3",
     "react": "16.13.1",
     "react-dom": "16.13.1",
