+++ pages/_app.tsx
@@ -1,8 +1,16 @@
-import '../styles/globals.css';
-import { AppProps } from 'next/app';
+import "../styles/globals.css";
+import { AppProps } from "next/app";
+import { useApollo } from "../lib/client";
+import { ApolloProvider } from "@apollo/client";
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  const apolloClient = useApollo(pageProps.initializeApollo);
+
+  return (
+    <ApolloProvider client={apolloClient}>
+      <Component {...pageProps} />
+    </ApolloProvider>
+  );
 }
 
 export default MyApp;
